+++ src/components/UriLabel/index.js
@@ -1,7 +1,8 @@
 // @flow
 import * as React from 'react';
 import { css, cx } from 'emotion';
-import { IconLink } from '../Icon';
+import { IconLink, IconBoldArrowRight } from '../Icon';
+import { colors } from '../../variables';
 import { Text } from '../Text';
 
 const styles = {
@@ -19,18 +20,37 @@ const styles = {
     text-overflow: ellipsis;
     white-space: nowrap;
     color: rgba(0, 0, 0, 0.65);
+  `,
+  hereIcon: css`
+    fill: ${colors.intentWarningAccent};
   `
 };
+
+const IconHere = ({
+  className,
+  ...props
+}) => (
+  <IconBoldArrowRight className={cx(styles.hereIcon, className)} {...props} />
+);
 
 type UriLabelProps = {
   className?: string,
   icon?: React.AbstractComponent<any>,
   title?: string,
   uri?: string,
+  weAreHere?: bool
 };
 
-export const UriLabel = ({ className, icon, title, uri }: UriLabelProps) => {
-  const Icon = icon || IconLink;
+export const UriLabel = (
+  {
+    className,
+    icon,
+    title,
+    uri,
+    weAreHere
+  }: UriLabelProps
+) => {
+  const Icon = icon || (weAreHere ? IconHere : IconLink);
 
   return (
     <div className={cx(styles.uriWrap, className)} title={title}>
@@ -38,4 +58,4 @@ export const UriLabel = ({ className, icon, title, uri }: UriLabelProps) => {
       <Text className={styles.uri} variant='p' tag='span'>{uri}</Text>
     </div>
   );
-}
+};