+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Added z-index value for notifications in global vars.
 
+- Added weAreHere prop to UriLabel.
+
 ## [0.29.0] - 2020-11-30
 
 - Hiding table head, when Table is empty.